# apps/api/openapi/generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/interviews/library": {
      "get": {
        "operationId": "InterviewsController_library",
        "summary": "List curated interview library entries (filterable).",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clause",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewLibraryListDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
...
    "/v1/interviews/{id}": {
      "get": {
        "operationId": "InterviewsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewsDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      },
      "patch": {
        "operationId": "InterviewsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewsDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      },
      "delete": {
        "operationId": "InterviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
...
    "/v1/peer-review/{id}": {
      "get": {
        "operationId": "PeerReviewController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "peer-review"
        ]
      },
      "patch": {
        "operationId": "PeerReviewController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "peer-review"
        ]
      },
      "delete": {
        "operationId": "PeerReviewController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "peer-review"
        ]
      }
    },
    "/v1/peer-review/{id}/comments": {
      "get": {
        "operationId": "PeerReviewController_listComments",
        "summary": "List comments on a peer-review package",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerReviewCommentListDto"
                }
              }
            }
          }
        },
        "tags": [
          "peer-review"
        ]
      },
      "post": {
        "operationId": "PeerReviewController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerReviewCommentDto"
                }
              }
            }
          }
        },
        "tags": [
          "peer-review"
        ]
      }
    },
    "/v1/peer-review/{id}/comments/{commentId}/resolve": {
      "post": {
        "operationId": "PeerReviewController_resolveComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerReviewCommentDto"
                }
              }
            }
          }
        },
        "tags": [
          "peer-review"
        ]
      }
    },
...
  },
  "info": {
    "title": "AuditForge ISO 42001 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InterviewsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firmId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "firmId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "InterviewLibraryEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "clauseRefs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicableModes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aiSystemClasses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "question": {
            "type": "string"
          },
          "followUps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "evidenceToSeek": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commonPitfalls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeBoxMinutes": {
            "type": "number"
          },
          "weight": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "role",
          "clauseRefs",
          "applicableModes",
          "aiSystemClasses",
          "question",
          "followUps",
          "evidenceToSeek",
          "commonPitfalls",
          "timeBoxMinutes",
          "weight"
        ]
      },
      "InterviewLibraryListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewLibraryEntryDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "PeerReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firmId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "firmId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PeerReviewCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "packageId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "object"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedBy": {
            "type": "string",
            "format": "uuid"
          },
          "resolutionNote": {
            "type": "string"
          },
          "flag": {
            "type": "string",
            "enum": [
              "standard",
              "security",
              "data-protection"
            ]
          }
        },
        "required": [
          "id",
          "packageId",
          "parentId",
          "authorId",
          "scope",
          "body",
          "createdAt",
          "flag"
        ]
      },
      "PeerReviewCommentListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeerReviewCommentDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}